# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  request.CreateUserRequest:
    properties:
      age:
        maximum: 60
        minimum: 18
        type: integer
      email:
        maxLength: 200
        minLength: 5
        type: string
      name:
        maxLength: 200
        minLength: 5
        type: string
      password:
        maxLength: 255
        minLength: 5
        type: string
      phone:
        maxLength: 15
        minLength: 10
        type: string
      role_id:
        maximum: 2
        minimum: 1
        type: integer
      username:
        maxLength: 200
        minLength: 5
        type: string
    required:
    - age
    - email
    - name
    - password
    - phone
    - role_id
    - username
    type: object
  request.LoginUserRequest:
    properties:
      password:
        maxLength: 255
        minLength: 5
        type: string
      username:
        maxLength: 200
        minLength: 5
        type: string
    required:
    - password
    - username
    type: object
...
  request.UpdateUserRequest:
    properties:
      age:
        maximum: 60
        minimum: 18
        type: integer
      email:
        maxLength: 200
        minLength: 5
        type: string
      name:
        description: ID    int    `validate:"required" json:"user_id"`
        maxLength: 200
        minLength: 5
        type: string
      phone:
        maxLength: 15
        minLength: 10
        type: string
    required:
    - age
    - email
    - name
    - phone
    type: object
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
      status:
        type: string
    type: object
host: localhost:8888
info:
  contact:
    email: rdbhagat999@gmail.com
    name: Ramandeep Bhagat
  description: A demo Service API in Golang using Gin Framework
  termsOfService: https://swagger.io/terms/
  title: Demo Service API
  version: 1.0.0
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login user in database
      parameters:
      - description: Login user
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/request.LoginUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Login user
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Logout user in database
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Logout user
  /auth/register:
    post:
      consumes:
      - application/json
      description: Save user in database
      operationId: CreateUser
      parameters:
      - description: Create user
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/request.CreateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Register user
...
  /users:
    get:
      consumes:
      - application/json
      description: Returns a list of users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get all users
  /users/{userId}:
    delete:
      consumes:
      - application/json
      description: Delete user from database
      parameters:
      - description: Delete user by id
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Delete user
    get:
      consumes:
      - application/json
      description: Returns a single user when userId maches id
      parameters:
      - description: Find user by id
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get a single user by its id
    put:
      consumes:
      - application/json
      description: Update and save user in database
      parameters:
      - description: Update user by id
        in: path
        name: userId
        required: true
        type: string
      - description: Update user
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/request.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Update user
  /users/authuser:
    get:
      consumes:
      - application/json
      description: Get user in database
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get authenticated user
swagger: "2.0"